Reorganized folders. Moved page level components to views.
